requirements.txt
aiobotocore==3.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
botocore==1.42.61
cachetools==7.0.3
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==43.0.3
cycler==0.12.1
databricks-sdk==0.96.0
debugpy==1.8.20
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.66.1
dvc-data==3.18.3
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-s3==3.3.0
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.2.1
fastapi==0.135.1
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gto==1.9.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
multidict==6.7.1
narwhals==2.18.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
oauth2client==4.1.3
omegaconf==2.3.0
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parso==0.8.6
pathspec==0.12.1
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
s3fs==2026.2.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.1
seaborn==0.13.2
semver==3.0.4
setuptools==82.0.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.48
sqlparse==0.5.5
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.52.1
streamlit==1.55.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
voluptuous==0.16.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.2
xgboost==3.2.0
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.0
